Handle non-breaking hyphens better.
In Trados, non-breaking hyphens were encoded as <ph>\_</ph> before.
This is now replaced by - in tmx2lfl.pl (see commit of that file in CVS repo).

In new versions of Trados, as well as in MS Word, non-breaking hyphens are
encoded as soft-hyphens.  This is a terrible choice, but we should support it
anyway, so now the TMX extraction changes the soft hyphen to - as well.