Factor out most corpus cleanup code from preparation/Makefile to clean-utf8-text.pl
Warning about multi-line embedded perl scripts inside Makefiles:

An earlier version of this patch had a bug in interaction between make and
multi-line embedded perl scripts.

When a Makefile contains a multi-line command, joined with \ escaping the
newline, GNU make does not strip that \ and newline, but rather passes it to
the shell for its interpretation.
http://www.gnu.org/software/make/manual/make.html#Splitting-Recipe-Lines

The consequence is that a multi-line embedded Perl script receives these \ as
part of the script. If the closing single quote is on its own line, that causes
a syntax error. My first hacky fix was to put a 1 before that closing single
quote, but frankly, I really don't like this. I'd be much happier with a
solution that stripped those newline-escaping \ before passing the script to
Perl.

Since the final script is short, my solution is to put the Perl script on one
line.

An alternative solution found by Darlene would be to define the script is a
variable first, e.g.:
SCRIPT='foo; \
 bar; \
 baz;'

and then embed that:
target:
	perl -e ${SCRIPT}

The rules for variable definition in GNU Make strip out the \ and newlines, so the
result is as desired:
	perl -e 'foo; bar; baz;'

Ref: http://www.gnu.org/software/make/manual/make.html#Splitting-Recipe-Lines